preliminary commit to refactor S3 reading logic
