+++ tools/test_history.py
@@ -6,11 +6,8 @@ import json
 import subprocess
 from collections import defaultdict
 from datetime import datetime
-from typing import Any, Dict, List, Optional, Set, Tuple, Union, cast
-
-import boto3  # type: ignore[import]
-import botocore  # type: ignore[import]
-from typing_extensions import Literal, TypedDict
+from typing import Any, Dict, List, Optional, Set, Tuple
+from torch.testing._internal.s3_stat_parser import (get_S3_bucket_readonly, get_cases, Report)
 
 
 def get_git_commit_history(
@@ -35,73 +32,6 @@ def get_object_summaries(*, bucket: Any, sha: str) -> Dict[str, List[Any]]:
         by_job[job].append(summary)
     return dict(by_job)
 
-
-# TODO: consolidate these typedefs with the identical ones in
-# torch/testing/_internal/print_test_stats.py
-
-Commit = str  # 40-digit SHA-1 hex string
-Status = Optional[Literal['errored', 'failed', 'skipped']]
-
-
-class CaseMeta(TypedDict):
-    seconds: float
-
-
-class Version1Case(CaseMeta):
-    name: str
-    errored: bool
-    failed: bool
-    skipped: bool
-
-
-class Version1Suite(TypedDict):
-    total_seconds: float
-    cases: List[Version1Case]
-
-
-class ReportMetaMeta(TypedDict):
-    build_pr: str
-    build_tag: str
-    build_sha1: Commit
-    build_branch: str
-    build_job: str
-    build_workflow_id: str
-
-
-class ReportMeta(ReportMetaMeta):
-    total_seconds: float
-
-
-class Version1Report(ReportMeta):
-    suites: Dict[str, Version1Suite]
-
-
-class Version2Case(CaseMeta):
-    status: Status
-
-
-class Version2Suite(TypedDict):
-    total_seconds: float
-    cases: Dict[str, Version2Case]
-
-
-class Version2File(TypedDict):
-    total_seconds: float
-    suites: Dict[str, Version2Suite]
-
-
-class VersionedReport(ReportMeta):
-    format_version: int
-
-
-# report: Version2Report implies report['format_version'] == 2
-class Version2Report(VersionedReport):
-    files: Dict[str, Version2File]
-
-
-Report = Union[Version1Report, VersionedReport]
-
-
 def get_jsons(
     jobs: Optional[List[str]],
     summaries: Dict[str, Any],
@@ -114,59 +44,6 @@ def get_jsons(
         job: json.loads(bz2.decompress(summaries[job].get()['Body'].read()))
         for job in keys
     }
-
-
-# TODO: consolidate this with the case_status function from
-# torch/testing/_internal/print_test_stats.py
-def case_status(case: Version1Case) -> Status:
-    for k in {'errored', 'failed', 'skipped'}:
-        if case[k]:  # type: ignore[misc]
-            return cast(Status, k)
-    return None
-
-
-# TODO: consolidate this with the newify_case function from
-# torch/testing/_internal/print_test_stats.py
-def newify_case(case: Version1Case) -> Version2Case:
-    return {
-        'seconds': case['seconds'],
-        'status': case_status(case),
-    }
-
-
-# TODO: consolidate this with the simplify function from
-# torch/testing/_internal/print_test_stats.py
-def get_cases(
-    *,
-    data: Report,
-    filename: Optional[str],
-    suite_name: Optional[str],
-    test_name: str,
-) -> List[Version2Case]:
-    cases: List[Version2Case] = []
-    if 'format_version' not in data:  # version 1 implicitly
-        v1report = cast(Version1Report, data)
-        suites = v1report['suites']
-        for sname, v1suite in suites.items():
-            if sname == suite_name or not suite_name:
-                for v1case in v1suite['cases']:
-                    if v1case['name'] == test_name:
-                        cases.append(newify_case(v1case))
-    else:
-        v_report = cast(VersionedReport, data)
-        version = v_report['format_version']
-        if version == 2:
-            v2report = cast(Version2Report, v_report)
-            for fname, v2file in v2report['files'].items():
-                if fname == filename or not filename:
-                    for sname, v2suite in v2file['suites'].items():
-                        if sname == suite_name or not suite_name:
-                            v2case = v2suite['cases'].get(test_name)
-                            if v2case:
-                                cases.append(v2case)
-        else:
-            raise RuntimeError(f'Unknown format version: {version}')
-    return cases
 
 
 def make_column(
@@ -455,9 +332,7 @@ indicated test was not found in that report.
         parser.error('No jobs specified.')
 
     commits = get_git_commit_history(path=args.pytorch, ref=args.ref)
-
-    s3 = boto3.resource("s3", config=botocore.config.Config(signature_version=botocore.UNSIGNED))
-    bucket = s3.Bucket('ossci-metrics')
+    bucket = get_S3_bucket_readonly('ossci-metrics')
 
     display_history(
         bucket=bucket,
